Move methods from PackingUtils into PackingPlanBuilder
