--- heron/packing/src/java/com/twitter/heron/packing/NullPacking.java
@@ -1,38 +1,0 @@
-// Copyright 2016 Twitter. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.twitter.heron.packing;
-
-import java.util.HashSet;
-
-import com.twitter.heron.api.generated.TopologyAPI;
-import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.packing.IPacking;
-import com.twitter.heron.spi.packing.PackingPlan;
-
-public class NullPacking implements IPacking {
-
-  @Override
-  public void initialize(Config config, TopologyAPI.Topology topology) {
-  }
-
-  @Override
-  public PackingPlan pack() {
-    return new PackingPlan("", new HashSet<PackingPlan.ContainerPlan>());
-  }
-
-  @Override
-  public void close() {
-  }
-}
