+++ heron/packing/src/java/com/twitter/heron/packing/PackingPlanBuilder.java
@@ -13,13 +13,17 @@
 // limitations under the License.
 package com.twitter.heron.packing;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeSet;
 import java.util.logging.Logger;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Optional;
 
 import com.twitter.heron.spi.packing.InstanceId;
@@ -110,7 +114,7 @@ public class PackingPlanBuilder {
   // build container plan sets by summing up instance resources
   public PackingPlan build() {
     assertResourceSettings();
-    Set<PackingPlan.ContainerPlan> containerPlans = PackingUtils.buildContainerPlans(
+    Set<PackingPlan.ContainerPlan> containerPlans = buildContainerPlans(
         this.containers, this.componentRamMap,
         this.defaultInstanceResource, this.requestedContainerPadding);
 
@@ -130,7 +134,7 @@ public class PackingPlanBuilder {
               new Container(this.maxContainerResource, this.requestedContainerPadding));
         }
       } else {
-        newContainerMap = PackingUtils.getContainers(
+        newContainerMap = getContainers(
             this.existingPacking, this.requestedContainerPadding);
       }
     }
@@ -156,5 +160,118 @@ public class PackingPlanBuilder {
       throw new PackingException(
           "maxContainerResource must be set on PackingPlanBuilder before modifying containers");
     }
+  }
+
+  /**
+   * Estimate the per instance and topology resources for the packing plan based on the ramMap,
+   * instance defaults and paddingPercentage.
+   *
+   * @return container plans
+   */
+  private static Set<PackingPlan.ContainerPlan> buildContainerPlans(
+      Map<Integer, Container> containerInstances,
+      Map<String, Long> ramMap,
+      Resource instanceDefaults,
+      double paddingPercentage) {
+    Set<PackingPlan.ContainerPlan> containerPlans = new HashSet<>();
+
+    for (Integer containerId : containerInstances.keySet()) {
+      Container container = containerInstances.get(containerId);
+      if (container.getInstances().size() == 0) {
+        continue;
+      }
+
+      long containerRam = 0;
+      long containerDiskInBytes = 0;
+      double containerCpu = 0;
+
+      // Calculate the resource required for single instance
+      Set<PackingPlan.InstancePlan> instancePlans = new HashSet<>();
+
+      for (PackingPlan.InstancePlan instancePlan : container.getInstances()) {
+        InstanceId instanceId = new InstanceId(instancePlan.getComponentName(),
+            instancePlan.getTaskId(), instancePlan.getComponentIndex());
+        long instanceRam;
+        if (ramMap.containsKey(instanceId.getComponentName())) {
+          instanceRam = ramMap.get(instanceId.getComponentName());
+        } else {
+          instanceRam = instanceDefaults.getRam();
+        }
+        containerRam += instanceRam;
+
+        // Currently not yet support disk or cpu config for different components,
+        // so just use the default value.
+        long instanceDisk = instanceDefaults.getDisk();
+        containerDiskInBytes += instanceDisk;
+
+        double instanceCpu = instanceDefaults.getCpu();
+        containerCpu += instanceCpu;
+
+        // Insert it into the map
+        instancePlans.add(new PackingPlan.InstancePlan(instanceId,
+            new Resource(instanceCpu, instanceRam, instanceDisk)));
+      }
+
+      containerCpu += (paddingPercentage * containerCpu) / 100;
+      containerRam += (paddingPercentage * containerRam) / 100;
+      containerDiskInBytes += (paddingPercentage * containerDiskInBytes) / 100;
+
+      Resource resource =
+          new Resource(Math.round(containerCpu), containerRam, containerDiskInBytes);
+
+      PackingPlan.ContainerPlan containerPlan =
+          new PackingPlan.ContainerPlan(containerId, instancePlans, resource);
+
+      containerPlans.add(containerPlan);
+    }
+
+    return containerPlans;
+  }
+
+  /**
+   * Sort the container plans based on the container Ids
+   *
+   * @return sorted array of container plans
+   */
+  public static PackingPlan.ContainerPlan[] sortOnContainerId(
+      Set<PackingPlan.ContainerPlan> containers) {
+    ArrayList<Integer> containerIds = new ArrayList<>();
+    PackingPlan.ContainerPlan[] currentContainers =
+        new PackingPlan.ContainerPlan[containers.size()];
+    for (PackingPlan.ContainerPlan container : containers) {
+      containerIds.add(container.getId());
+    }
+    Collections.sort(containerIds);
+    for (PackingPlan.ContainerPlan container : containers) {
+      int position = containerIds.indexOf(container.getId());
+      currentContainers[position] = container;
+    }
+    return currentContainers;
+  }
+
+  /**
+   * Generates the containers that correspond to the current packing plan
+   * along with their associated instances.
+   *
+   * @return Map of containers for the current packing plan, keyed by containerId
+   */
+  @VisibleForTesting
+  static Map<Integer, Container> getContainers(PackingPlan currentPackingPlan,
+                                               int paddingPercentage) {
+    Map<Integer, Container> containers = new HashMap<>();
+
+    //sort containers based on containerIds;
+    PackingPlan.ContainerPlan[] currentContainerPlans =
+        sortOnContainerId(currentPackingPlan.getContainers());
+
+    Resource capacity = currentPackingPlan.getMaxContainerResources();
+    for (PackingPlan.ContainerPlan currentContainerPlan : currentContainerPlans) {
+      Container container = new Container(capacity, paddingPercentage);
+      for (PackingPlan.InstancePlan instancePlan : currentContainerPlan.getInstances()) {
+        container.add(instancePlan);
+      }
+      containers.put(currentContainerPlan.getId(), container);
+    }
+    return containers;
   }
 }
