+++ heron/packing/src/java/com/twitter/heron/packing/PackingUtils.java
@@ -13,20 +13,12 @@
 // limitations under the License.
 package com.twitter.heron.packing;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.logging.Logger;
 
 import com.twitter.heron.api.generated.TopologyAPI;
 import com.twitter.heron.spi.common.Constants;
-import com.twitter.heron.spi.packing.InstanceId;
-import com.twitter.heron.spi.packing.PackingPlan;
 import com.twitter.heron.spi.packing.Resource;
 import com.twitter.heron.spi.utils.TopologyUtils;
 
@@ -46,10 +38,10 @@ public final class PackingUtils {
    * @param instanceResources The resources allocated to the instance
    * @return true if the instance is valid, false otherwise
    */
-  public static boolean isValidInstance(Resource instanceResources,
-                                        long minInstanceRam,
-                                        Resource maxContainerResources,
-                                        int paddingPercentage) {
+  private static boolean isValidInstance(Resource instanceResources,
+                                         long minInstanceRam,
+                                         Resource maxContainerResources,
+                                         int paddingPercentage) {
 
     if (instanceResources.getRam() < minInstanceRam) {
       LOG.severe(String.format(
@@ -106,110 +98,12 @@ public final class PackingUtils {
     return defaultInstanceResource.cloneWithRam(instanceRam);
   }
 
-  /**
-   * Estimate the per instance and topology resources for the packing plan based on the ramMap,
-   * instance defaults and paddingPercentage.
-   *
-   * @return container plans
-   */
-  public static Set<PackingPlan.ContainerPlan> buildContainerPlans(
-      Map<Integer, Container> containerInstances,
-      Map<String, Long> ramMap,
-      Resource instanceDefaults,
-      double paddingPercentage) {
-    Set<PackingPlan.ContainerPlan> containerPlans = new HashSet<>();
-
-    for (Integer containerId : containerInstances.keySet()) {
-      Container container = containerInstances.get(containerId);
-      if (container.getInstances().size() == 0) {
-        continue;
-      }
-
-      long containerRam = 0;
-      long containerDiskInBytes = 0;
-      double containerCpu = 0;
-
-      // Calculate the resource required for single instance
-      Set<PackingPlan.InstancePlan> instancePlans = new HashSet<>();
-
-      for (PackingPlan.InstancePlan instancePlan : container.getInstances()) {
-        InstanceId instanceId = new InstanceId(instancePlan.getComponentName(),
-            instancePlan.getTaskId(), instancePlan.getComponentIndex());
-        long instanceRam;
-        if (ramMap.containsKey(instanceId.getComponentName())) {
-          instanceRam = ramMap.get(instanceId.getComponentName());
-        } else {
-          instanceRam = instanceDefaults.getRam();
-        }
-        containerRam += instanceRam;
-
-        // Currently not yet support disk or cpu config for different components,
-        // so just use the default value.
-        long instanceDisk = instanceDefaults.getDisk();
-        containerDiskInBytes += instanceDisk;
-
-        double instanceCpu = instanceDefaults.getCpu();
-        containerCpu += instanceCpu;
-
-        // Insert it into the map
-        instancePlans.add(new PackingPlan.InstancePlan(instanceId,
-            new Resource(instanceCpu, instanceRam, instanceDisk)));
-      }
-
-      containerCpu += (paddingPercentage * containerCpu) / 100;
-      containerRam += (paddingPercentage * containerRam) / 100;
-      containerDiskInBytes += (paddingPercentage * containerDiskInBytes) / 100;
-
-      Resource resource =
-          new Resource(Math.round(containerCpu), containerRam, containerDiskInBytes);
-
-      PackingPlan.ContainerPlan containerPlan =
-          new PackingPlan.ContainerPlan(containerId, instancePlans, resource);
-
-      containerPlans.add(containerPlan);
-    }
-
-    return containerPlans;
-  }
-
-  /**
-   * Sort the container plans based on the container Ids
-   *
-   * @return sorted array of container plans
-   */
-  public static PackingPlan.ContainerPlan[] sortOnContainerId(
-      Set<PackingPlan.ContainerPlan> containers) {
-    ArrayList<Integer> containerIds = new ArrayList<>();
-    PackingPlan.ContainerPlan[] currentContainers =
-        new PackingPlan.ContainerPlan[containers.size()];
-    for (PackingPlan.ContainerPlan container : containers) {
-      containerIds.add(container.getId());
-    }
-    Collections.sort(containerIds);
-    for (PackingPlan.ContainerPlan container : containers) {
-      int position = containerIds.indexOf(container.getId());
-      currentContainers[position] = container;
-    }
-    return currentContainers;
-  }
-
   public static long increaseBy(long value, int paddingPercentage) {
     return value + (paddingPercentage * value) / 100;
   }
 
   public static double increaseBy(double value, int paddingPercentage) {
     return value + (paddingPercentage * value) / 100;
-  }
-
-  /**
-   * Allocate a new container of a given capacity
-   *
-   * @return the number of containers
-   */
-  public static int allocateNewContainer(List<Container> containers, Resource capacity,
-                                         int paddingPercentage) {
-    containers.add(new Container(capacity, paddingPercentage));
-    return containers.size();
   }
 
   /**
@@ -255,44 +149,6 @@ public final class PackingUtils {
       }
     }
     return new Resource(cpu, ram, disk);
-  }
-
-  /**
-   * Removes containers from tha allocation that do not contain any instances
-   */
-  public static void removeEmptyContainers(Map<Integer, List<InstanceId>> allocation) {
-    Iterator<Integer> containerIds = allocation.keySet().iterator();
-    while (containerIds.hasNext()) {
-      Integer containerId = containerIds.next();
-      if (allocation.get(containerId).isEmpty()) {
-        containerIds.remove();
-      }
-    }
-  }
-
-  /**
-   * Generates the containers that correspond to the current packing plan
-   * along with their associated instances.
-   *
-   * @return Map of containers for the current packing plan, keyed by containerId
-   */
-  static Map<Integer, Container> getContainers(PackingPlan currentPackingPlan,
-                                               int paddingPercentage) {
-    Map<Integer, Container> containers = new HashMap<>();
-
-    //sort containers based on containerIds;
-    PackingPlan.ContainerPlan[] currentContainerPlans =
-        PackingUtils.sortOnContainerId(currentPackingPlan.getContainers());
-
-    Resource capacity = currentPackingPlan.getMaxContainerResources();
-    for (PackingPlan.ContainerPlan currentContainerPlan : currentContainerPlans) {
-      Container container = new Container(capacity, paddingPercentage);
-      for (PackingPlan.InstancePlan instancePlan : currentContainerPlan.getInstances()) {
-        container.add(instancePlan);
-      }
-      containers.put(currentContainerPlan.getId(), container);
-    }
-    return containers;
   }
 
   public enum ScalingDirection {
