+++ heron/packing/tests/java/com/twitter/heron/packing/PackingUtilsTest.java
@@ -13,13 +13,8 @@
 // limitations under the License.
 package com.twitter.heron.packing;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -27,13 +22,9 @@ import org.junit.Test;
 import com.twitter.heron.api.generated.TopologyAPI;
 import com.twitter.heron.spi.common.ClusterDefaults;
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.common.Constants;
 import com.twitter.heron.spi.common.Context;
 import com.twitter.heron.spi.common.Keys;
-import com.twitter.heron.spi.packing.InstanceId;
-import com.twitter.heron.spi.packing.PackingPlan;
 import com.twitter.heron.spi.packing.Resource;
-import com.twitter.heron.spi.utils.PackingTestUtils;
 import com.twitter.heron.spi.utils.TopologyTests;
 
 public class PackingUtilsTest {
@@ -43,70 +34,6 @@ public class PackingUtilsTest {
       com.twitter.heron.api.Config topologyConfig) {
     return TopologyTests.createTopology("testTopology", topologyConfig, "spout", "bolt",
         spoutParallelism, boltParallelism);
-  }
-
-  private static Set<PackingPlan.ContainerPlan> generateContainers(Integer[] containerIds,
-                                                                   Integer[] instanceIds) {
-    Set<PackingPlan.ContainerPlan> containerPlan = new HashSet<>();
-    for (int containerId : containerIds) {
-      containerPlan.add(PackingTestUtils.testContainerPlan(containerId, instanceIds));
-    }
-    return containerPlan;
-  }
-
-  private static PackingPlan generatePacking(Map<Integer, List<InstanceId>> basePacking)
-      throws RuntimeException {
-    Resource resource = new Resource(2.0, 6 * Constants.GB, 25 * Constants.GB);
-
-    Set<PackingPlan.ContainerPlan> containerPlans = new HashSet<>();
-
-    for (int containerId : basePacking.keySet()) {
-      List<InstanceId> instanceList = basePacking.get(containerId);
-
-      Set<PackingPlan.InstancePlan> instancePlans = new HashSet<>();
-
-      for (InstanceId instanceId : instanceList) {
-        String componentName = instanceId.getComponentName();
-        Resource instanceResource;
-        switch (componentName) {
-          case "bolt":
-            instanceResource = new Resource(1.0, 2 * Constants.GB, 10 * Constants.GB);
-            break;
-          case "spout":
-            instanceResource = new Resource(1.0, 3 * Constants.GB, 10 * Constants.GB);
-            break;
-          default:
-            throw new RuntimeException(String.format("%s is not a valid component name",
-                componentName));
-        }
-        instancePlans.add(new PackingPlan.InstancePlan(instanceId, instanceResource));
-      }
-      PackingPlan.ContainerPlan containerPlan =
-          new PackingPlan.ContainerPlan(containerId, instancePlans, resource);
-
-      containerPlans.add(containerPlan);
-    }
-
-    return new PackingPlan("", containerPlans);
-  }
-
-
-  /**
-   * Tests the sorting of containers based on the container Id.
-   */
-  @Test
-  public void testContainerSortOnId() {
-
-    Integer[] containerIds = {5, 4, 1, 2, 3};
-    Integer[] instanceIds = {1, 2, 3};
-    Set<PackingPlan.ContainerPlan> containers = generateContainers(containerIds, instanceIds);
-
-    PackingPlan.ContainerPlan[] currentContainers = PackingUtils.sortOnContainerId(containers);
-
-    Assert.assertEquals(containerIds.length, currentContainers.length);
-    for (int i = 0; i < currentContainers.length; i++) {
-      Assert.assertEquals((currentContainers[i]).getId(), i + 1);
-    }
   }
 
   /**
@@ -132,18 +59,6 @@ public class PackingUtilsTest {
   }
 
   /**
-   * Tests containerAllocation method
-   */
-  @Test
-  public void testAllocateContainer() {
-    ArrayList<Container> containers = new ArrayList<>();
-    int expectedNumContainers = 1;
-    Assert.assertEquals(expectedNumContainers, PackingUtils.allocateNewContainer(containers,
-        new Resource(1, 1, 1), 10));
-  }
-
-
-  /**
    * Tests the component scale up and down methods.
    */
   @Test
@@ -164,59 +79,6 @@ public class PackingUtilsTest {
     Assert.assertEquals(2, componentToScaleDown.size());
     Assert.assertEquals(-2, (int) componentToScaleDown.get("spout"));
     Assert.assertEquals(-1, (int) componentToScaleDown.get("bolt2"));
-  }
-
-  /**
-   * Tests the component scale up and down methods.
-   */
-  @Test
-  public void testRemoveContainers() {
-
-    Map<Integer, List<InstanceId>> allocation = new HashMap<>();
-    allocation.put(1, new ArrayList<InstanceId>());
-
-    ArrayList<InstanceId> instances = new ArrayList<InstanceId>();
-    for (int i = 1; i <= 3; i++) {
-      instances.add(new InstanceId("instance", 1, i));
-    }
-    allocation.put(2, instances);
-
-    ArrayList<InstanceId> instances2 = new ArrayList<InstanceId>();
-    instances2.add(new InstanceId("instance", 2, 1));
-    allocation.put(3, instances2);
-
-    allocation.put(4, new ArrayList<InstanceId>());
-
-    PackingUtils.removeEmptyContainers(allocation);
-    Assert.assertEquals(2, allocation.size());
-    Assert.assertEquals(instances, allocation.get(2));
-    Assert.assertEquals(instances2, allocation.get(3));
-  }
-
-  /**
-   * Tests the getContainers method.
-   */
-  @Test
-  public void testGetContainers() {
-
-    int paddingPercentage = 10;
-    Map<Integer, List<InstanceId>> packing = new HashMap<>();
-    packing.put(7, Arrays.asList(
-        new InstanceId("spout", 1, 0),
-        new InstanceId("bolt", 2, 0)));
-    packing.put(3, Arrays.asList(
-        new InstanceId("spout", 3, 0),
-        new InstanceId("bolt", 4, 0)));
-
-    PackingPlan packingPlan = generatePacking(packing);
-    Map<Integer, Container> containers = PackingUtils.getContainers(packingPlan, paddingPercentage);
-    Assert.assertEquals(packing.size(), containers.size());
-    for (Integer containerId : packing.keySet()) {
-      Container foundContainer = containers.get(containerId);
-      Assert.assertEquals(paddingPercentage, foundContainer.getPaddingPercentage());
-      Assert.assertEquals(packingPlan.getMaxContainerResources(), foundContainer.getCapacity());
-      Assert.assertEquals(2, foundContainer.getInstances().size());
-    }
   }
 
   @Test
